Remove redundant or decorative references to the Director
A number of sentences in the Process made mentions of the Director that
did not add any information, for example by mentioning something like
“the Director’s decision to FOO” when FOO is already defined to be
something only the Director can decide.

This removes all such redundant mentions of the Director.

This commit does not change any of the Director’s powers.